pyproject.toml
[project]
name = "trackfm"
version = "0.1.0"
description = "Multi-Horizon Trajectory Forecasting with Horizon-Aware Transformers"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "Saul Pingerman"}]
urls = {repository = "https://github.com/saulpingerman/track-fm"}
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "matplotlib>=3.5.0",
    "polars>=0.20.0",
    "numpy>=1.21.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "tqdm>=4.65.0",
    "boto3>=1.34.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "ipykernel>=6.0.0",
]
